feat: refactor lessonItem consts
