--- apps/api/src/lessons/const/lessonItems.const.ts
@@ -1,3 +1,0 @@
-export const LESSON_ITEM_TYPES = ["text_block", "file", "question"];
-
-export type LessonItemType = (typeof LESSON_ITEM_TYPES)[number];
