+++ apps/api/src/lessons/lessonItems.type.ts
@@ -1,0 +1,3 @@
+export const LESSON_ITEM_TYPES = ["text_block", "file", "question"] as const;
+
+export type LessonItemType = (typeof LESSON_ITEM_TYPES)[number];
